notnormal/not_normal: fixed C throwing a type error from scipy.norm.ppf return, fixed initial threshold not being used in max cutoff determination, final iteration baseline and calculation trace now comes from the iteration which causes the first dip in Shapiro - this is counter-intuitive but the true best Shapiro lies between this iteration and the last - it is more consistent to take this iteration as the final (based on simulation experiment). This means the final events are synonymous with the event mask which causes the first dip (with a more accurate threshold).
notnormal/not_normal_gui: minor slider related bug fixes, minor change to the event seeking function, changing to the order that entries are flashes so that the after calls do not get buried in blocking computation.

notnormal/simulate: baseline generation is now in terms of nanopore or nano-impact (voltage, resistance, resistance change and Cottrell numerator) - selection is based on the chosen (optional) input parameters - capacitive transient not included as honestly so minor it is irrelevant, removal of the exponential window event function as it gives incredibly inconsistent results once added to the trace.