notnormal/not_normal: saved the initial threshold for improved maximum cutoff calculation. Removed adding of the original cutoff from calculating the cutoff as it is not additive (although, was an approximation). If the calculated cutoff is below that of the original cutoff used to detect, it is set at the original cutoff.
notnormal/not_normal_gui: ARM mac compatibility. Adjusted visuals in terms of entry text sizes and widths. Fixed some entries not allowing return to set the value. Threading of the initial estimate and iteration is now checked periodically with .after() as issues arose on ARM mac. Event vectors are not stored in base64 for the output csv (excel cell size limit).

notnormal/results: added a remove method to the Events dataclass based on event ID.

notnormal/simulate: added some functionality for dielectric noise types and an improved method for calculating the STD based on the PSD integral. This file is subjective to change a lot in the coming period.