notnormal/not_normal_gui: Added event overlay plot to the feature window, requires specific config; (hopefully) fixed some focusing oddities; feature window will now update based on what analysis was last ran and stay open with the same settings; feature window will now keep settings when closed for any specific run, and thus, reopen with the same settings; added the options for FWHM/FWQM feature extraction, this only impacts duration measurements. This is largely because of long-tailed events.
For a long time I've thought that a Gaussian window analogue of the methodology would perform more favourably, using the inverse error func to determine maximum cutoffs. It turns out, it performs almost exactly the same.

I also have thought that maybe the individual replacement segments to determine the local baseline should be reversed. This also makes little to no difference, and in any case, is not better.